moving fsaverage fetcher, formatting
