+++ nilearn/datasets/__init__.py
@@ -5,14 +5,14 @@ Helper functions to download NeuroImaging datasets
 from .struct import (fetch_icbm152_2009, load_mni152_template,
                      load_mni152_brain_mask, fetch_oasis_vbm,
                      fetch_icbm152_brain_gm_mask,
-                     MNI152_FILE_PATH)
+                     MNI152_FILE_PATH, fetch_surf_fsaverage5)
 from .func import (fetch_haxby_simple, fetch_haxby, fetch_nyu_rest,
                    fetch_adhd, fetch_miyawaki2008,
                    fetch_localizer_contrasts, fetch_abide_pcp,
                    fetch_localizer_button_task,
                    fetch_localizer_calculation_task, fetch_mixed_gambles,
                    fetch_megatrawls_netmats, fetch_cobre,
-                   fetch_surf_nki_enhanced, fetch_surf_fsaverage5)
+                   fetch_surf_nki_enhanced)
 from .atlas import (fetch_atlas_craddock_2012, fetch_atlas_destrieux_2009,
                     fetch_atlas_harvard_oxford, fetch_atlas_msdl,
                     fetch_coords_power_2011,
