+++ nilearn/datasets/description/fsaverage5.rst
@@ -3,24 +3,19 @@ fsaverage5
 
 Notes
 -----
-Fsaverage5 standard surface as distributed with Freesurfer (Dale et al, 1999,
-Fischl et al, 1999)
+Fsaverage5 standard surface as distributed with Freesurfer (Fischl et al, 1999)
 
 Content
 -------
-    :'pial_left': pial surface mesh for the left hemisphere
-    :'pial_right': pial surface mesh for the right hemisphere
-    :'infl_left': inflated pial surface mesh for the left hemisphere
-    :'infl_right': inflated pial surface mesh for the right hemisphere
-    :'sulc_left': sulcal depth data for the left hemisphere
-    :'sulc_right': sulcal depth data for the right hemisphere
+    :'pial_left': Gifti file, left hemisphere pial surface mesh
+    :'pial_right': Gifti file, right hemisphere pial surface mesh
+    :'infl_left': Gifti file, left hemisphere inflated pial surface mesh
+    :'infl_right': Gifti file, right hemisphere inflated pial
+                   surface mesh
+    :'sulc_left': Gifti file, left hemisphere sulcal depth data
+    :'sulc_right': Gifti file, right hemisphere sulcal depth data
 
 References
 ----------
-Dale et al, (1999). Cortical surface-based analysis.I. Segmentation and
-surface reconstruction. Neuroimage 9.
-http://dx.doi.org/10.1006/nimg.1998.0395
-
-Fischl et al, (1999). Cortical surface-based analysis. II: Inflation,
-flattening, and a surface-based coordinate system. Neuroimage 9.
-http://dx.doi.org/10.1006/nimg.1998.0396
+Fischl et al, (1999). High-resolution intersubject averaging and a
+coordinate system for the cortical surface. Hum Brain Mapp 8, 272-284.
