+++ nilearn/datasets/func.py
@@ -1718,16 +1718,16 @@ def fetch_surf_nki_enhanced(n_subjects=10, data_dir=None,
         By default, 10 subjects will be loaded. If None is given,
         all 102 subjects will be loaded.
 
-    data_dir: string, optional
+    data_dir: str, optional
         Path of the data directory. Used to force data storage in a specified
         location. Default: None
 
-    url: string, optional
+    url: str, optional
         Override download URL. Used for test only (or if you setup a mirror of
         the data). Default: None
 
     resume: bool, optional (default True)
-        If true, try resuming download if possible.
+        If True, try resuming download if possible.
 
     verbose: int, optional (default 1)
         Defines the level of verbosity of the output.
@@ -1843,97 +1843,4 @@ def fetch_surf_nki_enhanced(n_subjects=10, data_dir=None,
 
     return Bunch(func_left=func_left, func_right=func_right,
                  phenotypic=phenotypic,
-                 description=fdescr)
-
-
-def fetch_surf_fsaverage5(data_dir=None, url=None, resume=True, verbose=1):
-
-    """ Download Freesurfer fsaverage5 surface
-    Parameters
-    ----------
-    data_dir: string, optional
-        Path of the data directory. Used to force data storage in a specified
-        location. Default: None
-
-    url: string, optional
-        Override download URL. Used for test only (or if you setup a mirror of
-        the data). Default: None
-
-    resume: bool, optional (default True)
-        If true, try resuming download if possible.
-
-    verbose: int, optional (default 1)
-        Defines the level of verbosity of the output.
-
-    Returns
-    -------
-    data: sklearn.datasets.base.Bunch
-        Dictionary-like object, the interest attributes are :
-         - 'pial_left': pial surface mesh for the left hemisphere
-         - 'pial_right': pial surface mesh for the right hemisphere
-         - 'infl_left': inflated pial surface mesh for the left hemisphere
-         - 'infl_right': inflated pial surface mesh for the right hemisphere
-         - 'sulc_left': sulcal depth data for the left hemisphere
-         - 'sulc_right': sulcal depth data for the right hemisphere
-
-    References
-    ----------
-    Reuter et al, (2010). High-resolution intersubject averaging and a
-    coordinate system for the cortical surface. Hum Brain Mapp 8, 272-284.
-    URL http://dx.doi.org/10.1002/(sici)1097-0193(1999)8:4<272::aid-hbm10>3.0.
-    co;2-4
-    """
-    if url is None:
-        url = 'https://www.nitrc.org/frs/download.php/'
-
-    # Preliminary checks and declarations
-    dataset_name = 'fsaverage5'
-    data_dir = _get_dataset_dir(dataset_name, data_dir=data_dir,
-                                verbose=verbose)
-
-    # Dataset description
-    fdescr = _get_dataset_descr(dataset_name)
-
-    # Download fsaverage surfaces and sulcal information
-    surf_file = '%s.%s.gii'
-    surf_url = url + '%i/%s.%s.gii'
-    surf_nids = {'lh pial': 9344, 'rh pial': 9345,
-                 'lh infl': 9346, 'rh infl': 9347,
-                 'lh sulc': 9348, 'rh sulc': 9349}
-
-    pials = []
-    infls = []
-    sulcs = []
-    for hemi in [('lh', 'left'), ('rh', 'right')]:
-
-        pial = _fetch_files(data_dir,
-                            [(surf_file % ('pial', hemi[1]),
-                                surf_url % (surf_nids['%s pial' % hemi[0]],
-                                            'pial', hemi[1]),
-                              {})],
-                            resume=resume, verbose=verbose)
-        pials.append(pial)
-
-        infl = _fetch_files(data_dir,
-                            [(surf_file % ('pial_inflated', hemi[1]),
-                              surf_url % (surf_nids['%s infl' % hemi[0]],
-                                          'pial_inflated', hemi[1]),
-                              {})],
-                            resume=resume, verbose=verbose)
-        infls.append(infl)
-
-        sulc = _fetch_files(data_dir,
-                            [(surf_file % ('sulc', hemi[1]),
-                              surf_url % (surf_nids['%s sulc' % hemi[0]],
-                                          'sulc', hemi[1]),
-                              {})],
-                            resume=resume, verbose=verbose)
-        sulcs.append(sulc)
-
-    return Bunch(pial_left=pials[0],
-                 pial_right=pials[1],
-                 infl_left=infls[0],
-                 infl_right=infls[1],
-                 sulc_left=sulcs[0],
-                 sulc_right=sulcs[1],
                  description=fdescr)
